Fix Gravatar URL: use only pauseid@cpan.org address
Now only use the CPAN e-mail address (pauseid@cpan.org) because we are
using the gravatar URL to show the author's CPAN identity.
Using only the pauseid@cpan.org allows the author to use his Gravatar
account to assign his own image for his CPAN identity. The previous code
used the first email in the profile (which usually is the final
destination of the @cpan.org address) but the author may want to use it for
others purpose, so this may be a privacy leak.
By the way, this could make initial Gravatar slightly display faster as we
drop a Gravatar fallback.