Make the bug summaries less restrictive
Instead of having a summary entry have an arbitrary type, such as "rt", make it
point to the source of the aggregated data using a "source" attribute
containing, preferably, full URLs,

Additionally, don't *require* all sorts of different attributes to be supported
by every bug tracker we want to summarise. Just require the basics, but allow
for additional attributes to be provided.