Fixes initial file download issues for ratings script.
Fall back to stringification of the file path if the ratings script does
not already exist.  resolve() returns undef if the file does not already
exist, but in the case of a missing file, the object will still
stringify to the correct path. When called without resolve() in the
past, the latest version of the file would never be downloaded. So, this
should now cover the first case where the file does not exist as well as
the recurring case where the file has already been downloaded.