Use same prefix for --relative as for source/target
Found weird path mangling on local machine...
Discovered (by reading the git(1) code for diff.c)
that --relative just takes the length of the arg
and strips that many chars off the front of the paths.

Providing the same prefix for --relative is clearer and more logical
and in this case is much more likely to work.