Allow a BOM to precede the package declaration
There may be a way to use the file's encoding and not have to do it this way
but currently we're reading the file in binary mode and I don't want to toy
with changing that at the moment.  This should fix a few dists in the meantime.

Fixes #227.
Fixes cpan-api/metacpan-web#956.