Remove default path to fakepan for CPAN.
Having this default meant that indexing works on the VM and production
even if you haven't set your _local config.  The confusing part is that
you end up the API ends up looking in the Fakepan even if you indexed
tarballs in a completely different location, so source URLs end up
returning a 404 in most cases.  We already have a test config that is
pointing at fakepan, so best to force people to be explicit.