Make the tarball test work.
This is mostly about being able to instanciate a Tarball object at
all.  It reveals a bunch of problems that will be dealt with later.

The karma distribution used was too complicated, it needed a named pipe
and that has permissions problems.  Use something *waaay* simpler for right
now.

MetaCPAN::Model::Tarball needs a default tmpdir, but there's problems with
how it's storing the directory (it can't store a File::Temp object properly).

MetaCPAN::Model::Tarball needs to be a logger, so it has to be passed the
logger and level now.  This is sub-optimal, but it makes it work.