Terminology fix!  Tarball -> Archive.
CPAN releases are tarballs AND zipfiles (and some other weird things).
It's wrong to refer to them as "tarballs" and can lead to forgetting
about zip files.  Instead use "archive" or "archive file".

Also normalized the use of the "file" attribute to mean "the file I
will be operating on" between MC::Model::Release and
MC::Model::Archive.