Replace direct use of Archive::Any with MC::Model::Archive.
Then if we change how archives are extracted in the future it only has
to be done in one place.

MC::Model::Archive deliberately does not clean up a directory given to
it, we don't know the caller's intent, but it does use a temp
directory by default.  Someone can add an option later if it's needed.