rework set_first_release + add first script to tests
this change will prevent the race between setting all releases to
first=0 and then setting first=1 based on a query that depends
on previous put to finish on ES level.

it adds a run of MetaCPAN::Script::First to the setup test,
to ensure all releases are indexed correctly ('first'-wise)
before running the other tests.