merge the expanded and collapsed into a single sendpoint
Also merge the model methods into one entry point.

The frontend has previously munged the query, we take that munging now.
From that we determine if the results should be expanded or collapsed
based on the query itself.

For flexibility we now also all explicitly passing a flag indicating
whether we want collapsed results. Additionally we return in the data
whether the results are actually expanded or collapsed for inspection
which is especially useful when the backend decided what to do based on
the query.

Note that the result structure does not depend on collapsing or not,
just the data contained within it.