Provide backwards compatibility for clients which request only specific fields
We claim in our docs¹ that our convenience endpoints revert the newer
Elasticsearch behaviour of wrapping single values in arrays.  This makes
that so in more places, specifically the places where the client has
requested a limited subset of fields.

While this may seem like a pain for us and lead to uglier code, think
about it this way: either we do it for our API one time, or every client
that runs into the issue has to do it themselves many times over.

A better solution may exist which centralizes application of
single_valued_arrayref_to_scalar(), but this works for now.

See also GH#580 for an example case.²

This reverts a small change to tests which was made to accommodate the
new ES behaviour.

¹ https://github.com/metacpan/metacpan-examples/blob/master/README.md#upgrading-from-v0
² https://github.com/metacpan/metacpan-api/issues/580