The Fastly role is required to purge_author_key
It was throwing an error because after saving the profile because $self->purge_author_key didn't exist, which resulted in the frontend showing an empty profile with a success message. In the frontend the result $res came in with only a `raw` key, but no `error` key, so no error was ever displayed.

Fixes metacpan/metacpan-web#1846.