Created MetaCPAN::Query::*
This new namespace will be used to hold ES query code that
is not coupled with the model and will eventually replace
the model.

As a first step, moving the contents of MetaCPAN::Document::Favorite::Set
to the new namespace.
This will be used to serve a shared access for the autocomplete method.