fix archive test to account for variable generated files
The META files generated in the fake release archive will include the
version numbers of the modules used to generate them.  This means they
can have different content on different machines or different module
versions.  Use a regex on the content as a sanity check rather than a
SHA-1 sum for these files.