improve choices for interesting files
Rather than allow any files to match by its filename alone, split the
list in two.  One list of files that must be in the root directory
(mostly files with no extension, markdown files, or ALL CAPS files).
For pod or pm files, allow them to be in subdirectories.  Eliminate a
few top level directories that are commonly used for special cases and
not part of the dist metadata or module list.