new search end point with total collapsed entries
The existing search end points gives an array of arrays.  For collapsed
searches, each inner array represents a dist.  Because these arrays are
stored directly without a wrapping object, there is nowhere to store
additional data about the results.

For collapsed results, there is little reason to include all of the
inner file results.  For the metacpan front end, we only need the first
5 or so, as well as the count.  There isn't anywhere to store the count
without introducing a wrapping object for each search result.

The new end point returns an array of objects.  The objects include the
distribution name, the top five inner hits, and the total amount of
inner hits.