Canonize the path provided by META to match dir listing
When parsing a release we compare the paths of files in 'provides'
with the directory listing.
When these entries in META contain './' they don't match the listing.
This causes the 'modules' section in the 'release' entry to miss
files and in turn (if 'modules' ends up empty) the release cannot
be marked as 'latest'.