Script::Release - allow forcing authorized flag
In some cases we have to deal with incorrect data based on reindexing
after permissions have changed.
In those cases we may want to reindex (manually) while forcing the
check of authorized to 'true' knowing this was the case at the time
of the original release.

See GH#2117 (https://github.com/metacpan/metacpan-web/issues/2117)