add symlink from bin/api.pl back to former bin/queue.pl
As the queue app grew to become the bigger app that it now is, the name
didn't make sense anymore. Thus MetaCPAN::Queue became MetaCPAN::API the
script was renamed too. However for ease of migration, this symlink will
keep the existing puppet services able to run the minion workers that it
originally ran.

Once those are updated to point to bin/api.pl this symlink can then be
removed.