Add route for `/v1` path
At some point there will be an update to nginx that no longer strips the
`/v1` from the path. When this happens, having this mount point in place
will allow the fall through to the Catalyst app to continue.

This fixes an issue where the Mojolicous application is unable to
validate the specification for OpenAPI because the defined base path
does not exist.