Change to use metacpan-base
Instead of using a base perl docker image, use the new metacpan-base
image and build the api upon that.

The result should be faster build times and smaller images. Environment
variables have been introduced for being able to switch the server that
is being executed in the image. Those variables are set in the `.env`
file.