add files_by_category end point
The existing interesting_files end point had a long list of files to
include, with inconsistent rules for what file extensions and
capitalization would be allowed.  It also would just give the files back
in a single list.  The code for finding change logs partly reproduced
this.

Unify the logic for the files we search for, with consistent rules for
file extensions, capitalization, and handling of special dists like
"perl".  Keep the different types of files categorized so that
interesting_files can both filter the files it is querying for, and
adding a new end point files_by_category to easily find the files of a
specific type, like was previously possible for change logs.  Install
files and contributing docs are now listed on the front end, so this
should make it easier for them to find the files.