split examples into paths and files
The list of things we're trying to list for examples included 'ex' and
'examples'.  These don't both make sense to match in the same way.

Split the file and path matching. Match directories at the top level,
not just the path prefix. This prevents matching a /ext/ directory. And
match files starting with examples, but not ex.