Fix exception thrown when we try to find reverse deps on an unauthorized module
In this case a search for /v1/reverse_dependencies/dist/Math-Calc-Units
returned a 500

The reason is that this module is marked as latest, but also not marked
as authorized, presumably because it includes Parse::Yapp::Driver. So,
it does not meet the conditions of the query as set in
MetaCPAN::Query::Release::_get_latest_release().

_get_latest_release() was assuming success and returning a hashref with
no name and author rather than returning false. Because we did not
short-circuit here, a later query failed with "No value specified for
term query".

This will now return a 404:
{ "code" : 404, "message" : "The requested info could not be found" }