Apply coercion to allow round-trip of boolean through JSON
The field is_pause_custodial_account is declared as a Bool, but this
type restricst values to empty string, zero or one.  When served as
a JSON document, this gets converted to a JSON boolean 'true' or
'false'.  When parsed back to Perl, this becomes a JSON::Boolean
object.  JSON::Boolean objects do not match the type constraint of
Bool, and cannot be used to construct ..Document::Author.
In particular, this is breaking the ability to update the profiles
of metacpan in metacpan/metacpan-web/issues/2694

Adding `coerce => 1` allows the JSON::Boolean to get transformed to
'' or 1, permitting a round trip without changing any representation
of the boolean.