Revert "Revert "Bug fix: fix param order (GH#1126)""
This reverts commit 781f2141c08ccdb25e7ddc29a34940df9a3e0074.

The front end had been supplying the parameters incorrectly, and the API
was accepting them incorrectly. So the front end was displaying
correctly, but the API was incorrect. Fixing the API broke the
pagination, which led to the original revert. The front end has now been
fixed, so the API fix needs to be re-applied.