Only purge in latest script if making a change
The previous attempt to only purge changed dists in the latest script
tried to only mark for purging when marking for upgrade or downgrade.
But the if conditions to mark for upgrade/downgrade is not all of the
logic. There are later conditions to limit which dists will be upgraded.

Move the code marking dists to purge until after this later filtering.