fix digest calculation from misuse or tr
tr/// always transforms character classes. Trying to put a regex
character class wrapped in brackets will result in the brackets getting
transformed, and other characters getting converted incorrectly.

Fix the tr/// used to generate a digest for an ID. This is only used for
the contributors data, which is generally only updated when new releases
are indexed. When querying the data, the author and release names are
used, not the ID. The old records will continue to be found. If the
data is reindexed, an additional record may be created for a release,
but it should have the same data.

After making this change, eventually all of the contributor data should
be recalculated to make sure the IDs are correct, and the old records
purged. But they shouldn't interfere with anything until that is done.