tidy all files
The tests are currently only applied to a subset of files, but in the
future we will be using precious against all files.

Remove gitignore sorting from the tidyall config. It uses a different
sorting algorithm than what we will be using in the future with
precious. Once we fully transition to precious, the order will be
checked again. But for now it isn't a high priority.