remove index option from scripts and reimplement it for backup
The index option is meant to control the index that has all of our ES
data. But in the future, each document type will need to be in a
separate index. So the idea of having a global configurable index
doesn't make sense.

The backup script needs to be able to operate on all indexes, and
works generically with any type. Reimplement an index option for it,
allowing multiple indexes to be specified.