allow strings for backup script index option
Using command line options, the index option will always be assigned as
an arrayref. But the default was still a plain string, which would error
if no index option was given. The default could be changed to an array
ref, but it also makes sense to coerce a normal string.