Scripts: refresh all indices
Previously when trying to refresh indices, scripts would call
$self->index->refresh. This would refresh the "currently used" index.
That doesn't make any sense when splitting each type into its own index.
This was also using ElasticSearchX::Model, which we want to get rid of.

Instead, call ->indices->refresh via the Search::Elasticsearch object.
This will refresh all indices, which is fine for our purposes. In the
future, we could consider being more selective about which indices we
are refreshing, but this is no worse than the old behavior.