find es index/type via ESConfig rather than passing index around
Many parts of the code treated the index as the parent of all data, so
it was the thing being passed around. That will no be true in the
future.

Instead, ESConfig can give the path (index+type) of each named document
type. Convert most places passing around index to use es_doc_path.