Always use type coercion to construct Search::Elasticsearch object
Make all uses of elasticsearch_servers consistent, passing the value
directly to the MooseX::Types::ElasticSearch::ES type coercion. This
ensures we can use any value that the type will accept. This allows
adding additional parameters that will be passed to the constructor.

With that allowed, explicitly set the client version we want to use.
Since we are currently pinned to Search::Elasticsearch 2.03, this won't
have any direct impact. But it will assist in upgrading the
Search::Elasticsearch version.

A _local config with just a string specified will override the entire
config, which for now will continue to work fine. Eventually, those
should be updated to use a hash so they can override only the nodes.