remove term_vector and fielddata from pod.analyzed
As far as I can tell, these are unused. These fields are using the
string type, which will be automatically be converted to text or keyword
types. With term_vector and fielddata, this automatic conversion can't
be done. Removing them allow the same mapping to work on both
Elasticsearch 2.4 and 5.