update Search::Elasticsearch to latest
To support newer Elasticsearch versions, we want a new
Search::Elasticsearch version. It still supports older versions, as long
as the correct module is installed and it is instructed to use it.

We've already made the necessary changes to explicitly configure it to
use Search::Elasticsearch::Client::2_0::Direct, so we are now free to
upgrade.