pod is always analyzed
We don't need pod in both analyzed and non-analyzed form. We could
remove the pod.analyzed field, but it would be more complicated to
upgrade. Trying to index pod as non-analyzed with doc_values will
be a problem for large pod content.

Make both the main pod field and pod.analyzed analyzed, with doc_values
off. We can consider dropping the sub-field in the future.