include name and email in contributor index
All contributors, including those that can't be mapped to a PAUSE
author, should be included in the contributor index. This will allow the
web end points to look up all contributor information via the index,
rather than needing to re-calculate for each call.

Adds the name and email fields to the contributor index, and populates
them when calculating contributor data. The PAUSE id is no longer
required, so it can't be used for the id of the ES doc. Instead, we
delete all of the ids for a release before updating the data.
Previously, the contributors script would ignore existing data. Now, it
will delete and recreate the data.

For now, the release/contributors end point will still recalculate the
data for each call. After this change is deployed and all of the data is
updated, it can be updated to use the contributor index and be much
simpler. This will also reduce the code shared between indexing and the
API.