only load MooseX::Getopt in script role
MooseX::Getopt is used in our scripts, and we need to tell it about some
of the types we use. But that is only needed in the scripts. Rather than
setting it up in the type module, we can set it up in the role that is
used in every script. This means it won't be loaded in the web app.