use a default logger for module outside the scriprs
Log::Contextual requires a logger to be configured. The scripts confgure
one, but anything outside the scripts may not have one configured.
Log::Contextual::Easy::Default provides a default logger that warns if a
real logger isn't configured. This can allow the module to be used
outside the scripts without errors.