when searching for a package line, allow apostrophes in the module name
When searching for a package line to see if it is intentionally hidden,
we are matching against the specific module that we are using. But the
name that exists in the code could include apostrophes instead of double
colons, while the recorded module name is normalized to use colons.

Construct a regex to allow matching the module name but with apostrophe
package separators.