allow local testing config
It can be useful to modify the testing config, such as to point to an
alternate Elasticsearch server. When loading the testing config using a
modified config suffix, it will merge the data with the main config. But
this also prevents loading a local testing config.

Instead, we can use a custom config file name, and still allow a _local
suffix. Copy the extra config options from the main config to the test
config, since it will no longer be merging them.