match files to provides data based on exact match, not suffix
The files listed in provides data should be an exact match for the file
path we are using. The regex matching was added in
f3543c11ea543d8b9a688cfbc4c189b653c4b147 saying only "fixed regression".
I can barely guess at what it was supposedly fixing, but it was
definitely not the correct fix.