better check for previously extracted source files
When extracting source files for the API server, we always fix the
extracted path have a single directory with the same name as the
release. Since we are also using the directory as a staging area, we end
up with the release name doubled.

When checking for a previously extracted tarball, use the full path of
the directory, since old versions of the software or incomplete
extractions may have left the parent directory in place but not have the
full proper new paths.